Bump SpawnJS 2.1.7 -> 2.1.17 and AsyncFileSystem 2.1.1 -> 2.2.0
Standing preference: everything we touch runs the latest of our own libs.
Audited every SpawnDev pin against nuget.org, not just the one that was
noticed - RTC, RTC.Server, SpawnJS.Cryptography, BackgroundServices and the
three UnitTesting packages were already current; these two were not.

Checked the RESOLVED version rather than trusting the pin: project.assets.json
read 2.1.7 / 2.1.1 before (nothing was lifting them transitively) and reads
2.1.17 / 2.2.0 after, in both the library and the Demo.

Two real API breaks came with it, fixed rather than cast around:

FileSystemWritableFileStream.Seek and Truncate now take `long` instead of
`ulong` - the correct type, matching Stream. Five `(ulong)` casts removed
across AsyncFSFileStore and OpfsLayoutProbe.

IAsyncFS gained GetReadStream / GetWriteStream / OpenStream overloads taking
OPFSFileOptions. The NotABrowserFs test double implements them at their FULL
signature, and its old one-arg GetReadStream/GetWriteStream are gone rather
than left beside them: an extra one-arg overload on the concrete type wins
overload resolution over the interface member, which would have quietly
bypassed fileOptions at any call site using the short form.

Full PMT sweep: Failed 0, Passed 1023, Skipped 38 - identical to the
pre-bump baseline, so ten versions of SpawnJS moved with no behavioural
change here. That sweep is the point of this commit; the compiler cannot
see a behavioural regression across a dependency jump this size.

+++ SpawnDev.WebTorrent/Storage/AsyncFSFileStore.cs
@@ -284,7 +284,7 @@ public sealed class AsyncFSFileStore : IJSChunkStore
     {
         var path = PathOf(file);
         var writable = await GetWritableAsync(file, path).ConfigureAwait(false);
-        await writable.Seek((ulong)fileOffset).ConfigureAwait(false);
+        await writable.Seek(fileOffset).ConfigureAwait(false);
         await writable.Write(data).ConfigureAwait(false);
         _writesSinceCommit++;
         if (_writesSinceCommit >= WritesPerCommit) await CommitWritablesAsync().ConfigureAwait(false);
@@ -351,7 +351,7 @@ public sealed class AsyncFSFileStore : IJSChunkStore
             // The file is given its final length once, so an unwritten span reads as zeros at the right
             // OFFSET rather than shifting every later byte - see the note on GetSyncHandleAsync, which has
             // always done this. Truncate SETS the length, so repeating it would be a no-op anyway.
-            if (_lengthEstablished.Add(path)) await w.Truncate((ulong)file.Length).ConfigureAwait(false);
+            if (_lengthEstablished.Add(path)) await w.Truncate(file.Length).ConfigureAwait(false);
             _writables[path] = w;
             return w;
         }
